requirements.txt
mysql-connector-python
PyQt5
XlsxWriter